+++ spine-unity/Assets/Spine/Editor/spine-unity/Editor/Components/SkeletonRootMotionBaseInspector.cs
@@ -119,6 +119,20 @@ namespace Spine.Unity.Editor {
 			}
 
 			EditorGUILayout.PropertyField(rigidBody, rigidBodyLabel);
+			DisplayWarnings();
+		}
+
+		protected void DisplayWarnings () {
+			bool usesRigidbodyPhysics = rigidBody.objectReferenceValue != null || rigidBody2D.objectReferenceValue != null;
+			if (usesRigidbodyPhysics) {
+				var rootMotionComponent = (SkeletonRootMotionBase)serializedObject.targetObject;
+				var skeletonComponent = rootMotionComponent ? rootMotionComponent.TargetSkeletonAnimationComponent : null;
+				if (skeletonComponent != null && skeletonComponent.UpdateTiming == UpdateTiming.InUpdate) {
+					string warningMessage = "Skeleton component uses 'Advanced - Animation Update' mode 'In Update'.\n" +
+						"When using a Rigidbody, 'In FixedUpdate' is recommended instead.";
+					EditorGUILayout.HelpBox(warningMessage, MessageType.Warning, true);
+				}
+			}
 		}
 	}
 }